requirements.txt
aiogram==2.11.2
psycopg2==2.8.6
python-dateutil==2.8.1
scheduler-core>=1.0.0
python-telegram-bot-calendar==1.0.2